+++ main.py
@@ -3,11 +3,11 @@ from kivy.core.window import Window
 from kivymd.app import MDApp
 from kivymd.uix.label import MDLabel
 
-from components.video import SourceSelector, VideoDisplay
+from components.chooser import SourceSelector
+from components.video import VideoDisplay
 from components.hits import HitList
 from processor.model import Detector
 from kivymd.uix.boxlayout import MDBoxLayout
-from kivymd.uix.card import MDCard
 from kivymd.uix.card import MDCard
 from kivy.clock import Clock
 
@@ -38,19 +38,13 @@ class MainLayout(MDBoxLayout):
         stream_container = MDBoxLayout(orientation="vertical", size_hint=(0.75, 1))
 
         stream_card = MDCard(
-            orientation="vertical",
-            size_hint=(1, 1),
-            elevation=1,
-            radius=[0]
+            orientation="vertical", size_hint=(1, 1), elevation=1, radius=[0]
         )
 
         hits_container = MDBoxLayout(orientation="vertical", size_hint=(0.25, 1))
 
         hits_card = MDCard(
-            orientation="vertical",
-            size_hint=(1, 1),
-            elevation=1,
-            radius=[0]
+            orientation="vertical", size_hint=(1, 1), elevation=1, radius=[0]
         )
 
         stream_card.add_widget(video_display)
@@ -69,10 +63,7 @@ class R2App(MDApp):
     def build(self):
         self.theme_cls.theme_style = "Light"
         self.theme_cls.primary_palette = "BlueGray"
-        return MainLayout(
-            orientation="horizontal",
-            size_hint=(1, 1)
-        )
+        return MainLayout(orientation="horizontal", size_hint=(1, 1))
 
 
 if __name__ == "__main__":
